feat: added filter and search functionality
